backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
requests==2.26.0
python-dotenv==0.19.0
jinja2==3.0.1
aiofiles==0.7.0
python-multipart==0.0.5
pytz==2023.3
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1